blink_light_using_flask/requirements.txt
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pigpio==1.78
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
simple-websocket==1.0.0
urllib3==2.1.0
websocket-client==1.6.4
Werkzeug==3.0.1
wsproto==1.2.0
